# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BlastArena API
  version: 1.0.0
  description: |
    REST API for BlastArena, a multiplayer online grid-based explosive arena game.
    All endpoints are mounted under the `/api` prefix.
    Authentication uses JWT Bearer tokens (access token) and httpOnly cookies (refresh token).
servers:
  - url: /api
    description: API base path
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Registration disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username or email already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with credentials
      description: Returns access token in body and sets httpOnly refreshToken cookie.
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: httpOnly refreshToken cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log out (revoke refresh token)
      operationId: logout
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using cookie
      description: Uses the httpOnly refreshToken cookie. No Authorization header needed.
      security:
        - CookieAuth: []
      operationId: refreshToken
      responses:
        '200':
          description: Token refreshed
          headers:
            Set-Cookie:
              description: Updated httpOnly refreshToken cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: No refresh token or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/verify-email/{token}:
    get:
      tags: [Auth]
      summary: Verify email address
      security: []
      operationId: verifyEmail
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset email
      security: []
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Reset email sent (if account exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      security: []
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Invalid token or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ===================== User =====================
...
  /auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend email verification link
      operationId: resendVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Verification email sent (if email matches)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ===================== Local Co-Op =====================
  /local-coop/login:
    post:
      tags: [Local Co-Op]
      summary: Log in as local co-op player 2
      description: Authenticates P2 and sets an isolated httpOnly cookie. Requires P1 to be authenticated.
      operationId: localCoopLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocalCoopLoginRequest'
      responses:
        '200':
          description: P2 session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalCoopSession'
        '400':
          description: Invalid credentials or same user as P1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /local-coop/session:
    get:
      tags: [Local Co-Op]
      summary: Get current local co-op P2 session
      operationId: localCoopSession
      security: []
      responses:
        '200':
          description: Active P2 session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalCoopSession'
        '401':
          description: No session or invalid/expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /local-coop/socket-token:
    get:
      tags: [Local Co-Op]
      summary: Get a short-lived socket token for P2
      description: Converts the P2 httpOnly cookie into a JS-readable token (5min expiry) for socket auth.
      operationId: localCoopSocketToken
      responses:
        '200':
          description: Socket token for P2
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: No local co-op session or invalid/expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token passed in Authorization header
    CookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: httpOnly refresh token cookie (set by login/refresh endpoints)
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Message:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ---- Enums ----
    UserRole:
      type: string
      enum: [user, moderator, admin]
...
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 128
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/PublicUser'
        accessToken:
          type: string
        expiresIn:
          type: integer
          description: Token expiry in seconds
    PublicUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'

    # ---- User ----
...
    LocalCoopLoginRequest:
      type: object
      required: [username, password, duration]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 20
        password:
          type: string
          minLength: 1
          maxLength: 128
        duration:
          type: integer
          description: Session duration in hours (0 = session-only cookie)
          enum: [0, 1, 6, 12, 24]
    LocalCoopSession:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
        cosmetics:
          type: object
          description: Player cosmetic data for rendering
          additionalProperties: true

    # ---- User (new) ----
...